kinopoisk_tests/requirements.txt
selenium
requests
pytest
allure-pytest
pytest-html
